requirements.txt
wordcloud==1.9.2
sqlitedict==2.1.0
matplotlib==3.7.2
pandas==2.0.3
streamlit==1.28.0
pillow==10.1.0
streamlit-autorefresh==1.0.1